Implement initrd to x86_64 kernel and UI changes made next release
Initrd kernel module has been added to x86_64, previously the kernel was
depeneds on disk controller boot driver, which seems failing on some machine
or virtualization platform.
UI changes are on the way for next x86_64 virtualization release.

Manas Kamal Choudhury,
Xeneva Pvt Ltd.

+++ Kernel/string.cpp
@@ -38,7 +38,7 @@ typedef size_t WT;
 
 extern "C" void _fastcpy(void* dest, void* src, size_t count);
 
-void memset(void *targ, uint8_t val, uint32_t len){
+void memset(void *targ, int val, uint32_t len){
 	uint8_t *t = (uint8_t*)targ;
 	while (len--)
 		*t++ = val;
